fix tests, update some of README
