+++ lib/index.js
@@ -68,13 +68,12 @@ resolver.query('example.com', 'A')
   query(qname, qtype='A', method='POST') {
     return new Promise((resolve, reject) => {
       if (!isMethodAllowed(method)) {
-        reject(new MethodNotAllowedError(`Request method ${method} not allowed. Must be either 'GET' or 'POST'`))
-      } else {
-        let dnsMessage = makeQuery(qname, qtype);
-        sendDohMsg(dnsMessage, this.nameserver_url, method)
-          .then(resolve)
-          .catch(reject)
+        return reject(new MethodNotAllowedError(`Request method ${method} not allowed. Must be either 'GET' or 'POST'`))
       }
+      let dnsMessage = makeQuery(qname, qtype);
+      sendDohMsg(dnsMessage, this.nameserver_url, method)
+        .then(resolve)
+        .catch(reject)
     });
   }
 }
@@ -135,31 +134,31 @@ sendDohMsg(msg, url, method)
   .catch(console.error);
  */
 function sendDohMsg(packet, url, method) {
-  const buf = dnsPacket.encode(packet);
-  let requestOptions;
-  let headers = {};
-  if (method === 'POST') {
-    headers = {
-      'Content-Type': 'application/dns-message',
-      'Accept': 'application/dns-message',
-      'Content-Length': buf.length
-    };
-  } else if (method === 'GET') {
-    const dnsQueryParam = buf.toString('base64').toString('utf-8').replace(/=/g, '');
-    url = `${url}?dns=${dnsQueryParam}`;
-    headers = {
-      'Accept': 'application/dns-message',
-    };
-  }
-  url = new URL(url);
-  requestOptions = {
-    method: method,
-    hostname: url.hostname,
-    port: url.port || 443,
-    path: url.pathname + url.search,
-    headers: headers
-  };
   return new Promise((resolve, reject) => {
+    const buf = dnsPacket.encode(packet);
+    let requestOptions;
+    let headers = {};
+    if (method === 'POST') {
+      headers = {
+        'Content-Type': 'application/dns-message',
+        'Accept': 'application/dns-message',
+        'Content-Length': buf.length
+      };
+    } else if (method === 'GET') {
+      const dnsQueryParam = buf.toString('base64').toString('utf-8').replace(/=/g, '');
+      url = `${url}?dns=${dnsQueryParam}`;
+      headers = {
+        'Accept': 'application/dns-message',
+      };
+    }
+    url = new URL(url);
+    requestOptions = {
+      method: method,
+      hostname: url.hostname,
+      port: url.port || 443,
+      path: url.pathname + url.search,
+      headers: headers
+    };
     const request = https.request(requestOptions, (response) => {
       response.on('data', (d) => {
         const decoded = dnsPacket.decode(d);
@@ -167,7 +166,8 @@ function sendDohMsg(packet, url, method) {
       })
     });
     request.on('error', (err) => {
-      reject(err);
+      request.abort();
+      return reject(err);
     });
     if (method === 'POST') {
       request.write(buf)
@@ -191,82 +191,7 @@ function getRandomInt(min, max) {
   return Math.floor(Math.random() * (max - min + 1)) + min
 }
 
-/**
- * Default DoH lookup options
- *
- * @type {{dnssecOk: boolean, method: string, qtype: string, qname: string, ecsAddress: string, noRecursion: boolean, ecsSourcePrefixLength: number, url: null}}
- */
-const defaultDohLookupOptions = {
-  url: null,
-  qtype: 'A',
-  qname: '.',
-  noRecursion: false,
-  dnssecOk: false,
-  method: 'POST',
-  ecsAddress: '0.0.0.0',
-  ecsSourcePrefixLength: 0,
-};
-
-/**
- * Perform a DNS over HTTPS lookup given options
- *
- * See `defaultDohLookupOptions` above for list of options
- * and their defaults
- *
- * @param options {object} DNS lookup options
- * @returns {Promise<dnsPacket>}
- */
-function dohLookup(options) {
-  var url = options.url || defaultDohLookupOptions.url;
-  if (!url) {
-    throw new Error('Must provide a URL to send DoH lookup to');
-  }
-  const flags = options.noRecursion ? 0 : dnsPacket.RECURSION_DESIRED;
-  const type = options.qtype || defaultDohLookupOptions.qtype;
-  const name = options.qname || defaultDohLookupOptions.qname;
-  const method = options.method || defaultDohLookupOptions.method;
-  let dnsMessage = {
-    type: 'query',
-    id: getRandomInt(1, 65534),
-    flags: flags,
-    questions: [{
-      type: type,
-      name: name,
-    }]
-  };
-  if (options.dnssecOk) {
-    dnsMessage.additionals = [{
-      type: 'OPT',
-      name: '.',
-      udpPayloadSize: 4096,
-      flags: dnsPacket.DNSSEC_OK,
-      options: []
-    }]
-  }
-  if (options.ecsAddress) {
-    const family = options.ecsAddress.indexOf(':') !== -1 ? 2 : 1;
-    const sourcePrefixLength = options.sourcePrefixLength || defaultDohLookupOptions.sourcePrefixLength;
-    if (!dnsMessage.additionals || !dnsMessage.additionals.length) {
-      dnsMessage.additionals = [{
-        type: 'OPT',
-        name: '.',
-        udpPayloadSize: 4096,
-        options: []
-      }]
-    }
-    dnsMessage.additionals[0].options.push({
-      code: 8,
-      family: family,
-      sourcePrefixLength: sourcePrefixLength,
-      scopePrefixLength: 0,
-      ip: options.ecsAddress
-    })
-  }
-  return sendDohMsg(dnsMessage, url, method);
-}
-
 const dohjs = {
-  lookup: dohLookup,
   sendDohMsg: sendDohMsg,
   DohResolver: DohResolver,
   makeQuery: makeQuery,
