+++ lib/index.spec.js
@@ -19,7 +19,7 @@ test('DNS query message for example.com matches expected', () => {
     type: 'query',
     id: 0,
     flags: 256,
-  questions: [ { type: 'A', name: 'example.com' } ]});
+    questions: [ { type: 'A', name: 'example.com' } ]});
 });
 
 test('DohResolver should be created', () => {
@@ -59,4 +59,8 @@ test('sendDohMsg() works (and the example.com zone still has an A record)', () =
       expect(response.answers.length).toBeGreaterThanOrEqual(1);
     }
   )
+});
+
+afterAll(async done => {
+  done();
 });
